[CI] Fix concurrency group name for PR/push workflow
head_ref is only available on PRs. Without this change all pushes ended up
being in one group. On upstream it cancels all consecutive main pushes
(which is a little undesired), but on forks it canceled all consecutive
pushes (which is undesired if you want to test multiple different branches).