Fix the test_valgrind.sh script running tests under valgrind
The current script just checks if there is the
"ERROR SUMMARY: 0 errors from 0 contexts"
line in the output of a test.

It turned out that there can be many lines starting with
"ERROR SUMMARY:" in the output of one test - some with
errors and some without, so the current approach is wrong.
We have to check if there are no other lines starting with
"ERROR SUMMARY:" than this:
"ERROR SUMMARY: 0 errors from 0 contexts" instead.

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>