Do not fail linear allocator with OOM
The linear base allocator fails with OOM now,
when it runs out of memory requested in umf_ba_linear_create().
The linear base allocator is used to allocate structures
which size depends on number of nodes for example,
so it is not known before calling umf_ba_linear_create().

This patch adds a feature that the pool of linear allocator can "grow"
(a new memory region is allocated from OS in fact)
when more space is needed.

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>