Disable CMake build type checks on Windows
Do not rely on the CMAKE_BUILD_TYPE configuration variable in Windows
builds. This variable is ignored by MSVC, the build type is determined
during the build time based on the '--config' parameter.

There seems to be no way for disabling the proxy library build based on
the used config, so this change provides just notification messages on
all Windows builds and builds proxy library always.