Use PROJECT_VERSION not CMAKE_PROJECT_VERSION
When building as a subproject the `CMAKE_PROJECT_VERSION` is the root
project version not the current project version. As such, in a
subproject which makes use of UMF, such as UR, the `.so` or `.dll`
library version would be set to the UR project version, not the UMF
version.

This patch fixes this erroneous versioning by using `PROJECT_VERSION`
(and derivatives) in place of `CMAKE_PROJECT_VERSION`.