Fix hwloc patch application on windows
Windows `cmd.exe` doesn't have a `true` command, `(exit 0)` will do the
same thing but it will work on both linux and windows.

This issue only shows up on Windows and not on a first build, since when
building the first time the patch applies cleanly so the `||` branch is
not taken, but then when rebuilding the patch doens't apply and the
command errors out in the `||` branch.