Fix includes in memspace test fixtures
This file uses `std::find` and `std::any_of` which are part of the
`<algorithm>` header, but it wasn't including it.

When building with gcc 14 this caused build issues such as:

```
memspace_fixtures.hpp:204:26: error: ‘any_of’ is not a member of ‘std’
```

With this patch adding the include building UMF with gcc 14 works fine.