Wrap linker flags on Windows for IntelLLVM
The Intel C++ compiler requires linker flags to be wrapped, because
CMake passes them through the compiler driver.
Prefix Linker options with `LINKER:`.
CMake will transform it to the appropriate flag for the compiler driver:
(Nothing for MSVC, clang-cl and /Qoption,link for ICX), so this will
work for the other compilers too, and for earlier versions of CMake.

Fixes: #757
Signed-off-by: Gergely Meszaros <gergely.meszaros@intel.com>