Add Coarse provider
Add Coarse provider - a memory provider that can manage a memory
(i.e. handle the alloc() and free() ops) of a given pre-allocated
buffer or of an additional upstream provider (e.g. OS Memory Provider
or providers that do not support the free() operation,
like the file memory provider and the DevDax memory provider).

Co-developed-by: Rafał Rudnicki <rafal.rudnicki@intel.com>
Co-developed-by: Lukasz Dorau <lukasz.dorau@intel.com>

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>