Fix build with hwloc support disabled
`umfMemspaceCreateFromNumaArray` is defined in
`memspaces/memspace_numa.c` which is only added to the build if hwloc is
enabled.

When hwloc is enabled this is added as an optional symbol to the linker
script through `UMF_OPTIONAL_SYMBOLS_LINUX`.

When hwloc is disabled this symbol isn't available so this linker script
would fail, I believe the correct fix is just to remove the symbol from
the linker script.