Add optional coarse provider to umfPoolTest fixture
Add optional coarse provider to umfPoolTest fixture.
Add 5th argument (coarse_memory_provider_params_t *)
to the poolCreateExtParams tuple (can be nullptr).
If it is non-nullptr, coarse provider is created
with the given provider as the upstream provider.

It will be used to test the following providers:
- file provider
- devdax provider
as the upstream providers with the coarse provider
and different pool managers.

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>