Fix file_alloc() and file_open/close_ipc_handle()
Alignment passed to `file_alloc()` must be a power of two
and a multiple or a divider of the page size.
Align up the alignment in `file_alloc()` to the page size.

file_open_ipc_handle() has to use the memory protection
and visibility got from the IPC handle, so let's add
the memory protection and visibility to the IPC handle.

Ref: #838

Fixes: #848

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>