Fix output directory on Windows with non-VS generators
On Windows when configuring CMake with the Ninja generator and
`-DUMF_BUILD_SHARED_LIBRARY=ON` the `umf.dll` is output to the
`${CMAKE_BINARY_DIR}/bin/$<CONFIG>` directory, this is
problematic as it breaks tools like `urinfo.exe` and `sycl-ls.exe` in
development builds as they rely on `umf.dll` residing in the same
directory in order to be loaded.

This behavior is desirable when using Visual Studio generators, however.
Therefore, this patch changes the logic to check if `CMAKE_GENERATOR`
matches the `"Visual Studio"` string before appending `$<CONFIG>` to the
output directory.