Fix disjoint_pool unit tests
The UMF tests should not use the default system allocator
to allocate memory, because they may not work correctly
under the UMF proxy library when they and the proxy library
are dynamically linked with the same libumf.so file.

In such case the same tracker is used in both:
the test and the proxy library, so there may be conflicts
of memory allocations. An allocation made originally
by the the default system allocator in the test
is first added to the tracker by the proxy library
(which replaced the system allocator) and then the test
itself tries to add the same pointer to the same tracker
in the tracking provider of the pool that had made
this allocation.

Fixes: #240

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>