Simplify and improve trackingAllocationMerge()
We do not need to allocate a new mergedValue,
update critnib using critnib_insert() in the update mode
and free the old value. It is enough to just update atomically
the size of the first part:

utils_atomic_store_release_u64((uint64_t *)&lowValue->size, totalSize);

The tracker entries cannot be merged if they are used
(if they have children).

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>