Refactor coarse_alloc()
Refactor coarse_alloc(): move the code responsible for handling
the `(curr == NULL)` case just after `curr = find_free_block()`.
Add coarse_add_free_block().

Now (before this change) when a user calls `coarse_alloc(1, 2MB)`
(allocate 1 byte with 2MB alignment) and there is no suitable blok
inside the coarse allocator, a new block is allocated from the provider
of size `coarse->page_size` at least, but it is added to the inner lists
with the requested size of only 1 byte and the rest of space
will not be able to be utilized.

After this change, when a user calls `coarse_alloc(1, 2MB)`
(allocate 1 byte with 2MB alignment) and there is no suitable blok
inside the coarse allocator, a new block is allocated from the provider
of size `alignment` (>= `coarse->page_size`) and it is added to the inner
lists with the `alignment` size. The new free block will be used
as a whole or split and the rest of space can be utilized later.

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>