[CI] Run DocsBuild only when needed
Oftentimes pushes to branches and PRs doesn't change anything in docs/
or include/ directories which influence the documentation build. This
job takes ca. 1 minute to run with each push, so it's optimal to run it
only when relevant changes were made.

Also, there is no need for other pr_push jobs to wait for docs to build.
Developer can work on the docs and project build/tests in any order.