Check global state destruction in destructors
In some cases the global state may have been destroyed when we reach
`umfDestroyPool` or `umfDestroyProvider`, in which case actually going
through with the destruction will cause segmentation faults.

The way the global state is managed should maybe be re-worked, but this
should be an okay workaround to avoid segmentation faults until then.