Build dockers
I added a new Docker image for Ubuntu 24.04 and updated the others.
I also created two new workflows that check for changes in the Docker files.
If any changes are detected, the relevant Docker image will be rebuilt and pushed (unless it's a pull request).
The rebuilt images will be available after the merge.