Add support for versioning in git archives
While releasing UMF we should update VERSION file with the current version.
It will be used only if git is not available or tags were not fetched.

To underline it is not a "detailed version" we will add "-dev" in CMake version
parsing, so it will stand out (comparing to e.g. versions like "1.0.3-dev1-...").